chore: separate out different bb commands into files
